+++ tests/test_amqp.py
@@ -75,7 +75,10 @@ async def test_bad_command(amqp_client, amqp_actor):
     cmd = await amqp_client.send_command("amqp_actor", "bad_command")
     await cmd
 
-    assert "Command 'bad_command' failed." in cmd.replies[-1].message["error"]
+    assert (
+        "Command 'bad_command' does not exist or cannot be parsed."
+        in cmd.replies[-1].message["error"]
+    )
 
 
 async def test_send_command_actor_not_connected(amqp_client, amqp_actor):